Refactor auto benchmark unit tests
